+++ fitty_frontend_app/lib/base_app_bar.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
+
+class BaseAppBar extends StatelessWidget implements PreferredSizeWidget {
+  const BaseAppBar(Key? key) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      leading: IconButton(
+        icon: Icon(Icons.notifications),
+        onPressed: () {
+          // Handle the notification button tap here.
+        },
+      ),
+      actions: <Widget>[
+        IconButton(
+          icon: Icon(Icons.settings),
+          onPressed: () {
+            // Handle the settings button tap here.
+          },
+        ),
+      ],
+      bottom: TabBar(
+        tabs: [
+          Tab(icon: Icon(Icons.directions_car)),
+          Tab(icon: Icon(Icons.directions_transit)),
+          Tab(icon: Icon(Icons.directions_bike)),
+        ],
+      ),
+      title: Center(child: Text('FITTY')),
+    );
+  }
+
+  @override
+  // TODO: implement preferredSize
+  Size get preferredSize => Size.zero;
+}
